pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "easyocr>=1.7.2",
    "fastapi>=0.115.11",
    "loguru>=0.7.3",
    "openai>=1.65.2",
    "pillow>=11.1.0",
    "pinecone[grpc]>=6.0.1",
    "pymupdf>=1.25.3",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.38",
    "uvicorn[standard]>=0.34.0",
]

[dependency-groups]
dev = [
    "pyright>=1.1.396",
    "ruff>=0.9.9",
    "watchfiles>=1.0.4",
]

[tool.ruff]
include = ["pyproject.toml", "src/**/*.py"]
target-version = "py313"

[tool.ruff.lint]
extend-select = ["UP", "FAST", "PT", "FURB", "RUF006", "RUF012", "RUF015", "RUF100", "ASYNC", "C4", "DTZ", "B", "FBT"]
ignore = ["E731", "PT018", "B905", "FBT001", "FBT002"]
